Handle race conditions with some file sync services.
With some file sync services (observed with Nextcloud), a file can
disappear momentarily when being updated.  This causes qtodotxt2
to abort an auto-reload with an error, at which point any update can
cause external changes to be undone.

This patch introduces a slight delay between when a file update
is noticed and when auto-reload happens, which prevents this problem
in nearly all cases.